+++ HARK/ConsumptionSaving/ConsMedModel.py
@@ -7,7 +7,7 @@ from builtins import str
 from builtins import range
 import numpy as np
 from scipy.optimize import brentq
-from HARK import  AgentType, HARKobject, makeOnePeriodOOSolver
+from HARK import  AgentType, MetricObject, makeOnePeriodOOSolver
 from HARK.distribution import addDiscreteOutcomeConstantMean, Lognormal
 from HARK.utilities import (
     CRRAutilityP_inv,
@@ -56,7 +56,7 @@ utility_invP = CRRAutility_invP
 utilityPP = CRRAutilityPP
 
 
-class MedShockPolicyFunc(HARKobject):
+class MedShockPolicyFunc(MetricObject):
     """
     Class for representing the policy function in the medical shocks model: opt-
     imal consumption and medical care for given market resources, permanent income,
@@ -273,7 +273,7 @@ class MedShockPolicyFunc(HARKobject):
         return dcdShk, dMeddShk
 
 
-class cThruXfunc(HARKobject):
+class cThruXfunc(MetricObject):
     """
     Class for representing consumption function derived from total expenditure
     and consumption.
@@ -399,7 +399,7 @@ class cThruXfunc(HARKobject):
         return dcdShk
 
 
-class MedThruXfunc(HARKobject):
+class MedThruXfunc(MetricObject):
     """
     Class for representing medical care function derived from total expenditure
     and consumption.
@@ -942,24 +942,26 @@ class ConsMedShockSolver(ConsGenIncProcessSolver):
         vFuncBool,
         CubicBool,
     ):
-        self.assignParameters(
-            solution_next=solution_next,
-            IncShkDstn=IncShkDstn,
-            MedShkDstn=MedShkDstn,
-            LivPrb=LivPrb,
-            DiscFac=DiscFac,
-            CRRA=CRRA,
-            CRRAmed=CRRAmed,
-            Rfree=Rfree,
-            MedPrice=MedPrice,
-            pLvlNextFunc=pLvlNextFunc,
-            BoroCnstArt=BoroCnstArt,
-            aXtraGrid=aXtraGrid,
-            pLvlGrid=pLvlGrid,
-            vFuncBool=vFuncBool,
-            CubicBool=CubicBool,
-            PermGroFac=0.0,
-        )  # dummy value required?
+        """
+        Constructor for a new solver for a one period problem with idiosyncratic
+        shocks to permanent and transitory income and shocks to medical need.
+        """
+        self.solution_next = solution_next
+        self.IncShkDstn = IncShkDstn
+        self.MedShkDstn = MedShkDstn
+        self.LivPrb = LivPrb
+        self.DiscFac = DiscFac
+        self.CRRA = CRRA
+        self.CRRAmed = CRRAmed
+        self.Rfree = Rfree
+        self.MedPrice = MedPrice
+        self.pLvlNextFunc = pLvlNextFunc
+        self.BoroCnstArt = BoroCnstArt
+        self.aXtraGrid = aXtraGrid
+        self.pLvlGrid = pLvlGrid
+        self.vFuncBool = vFuncBool
+        self.CubicBool = CubicBool
+        self.PermGroFac = 0.0
         self.defUtilityFuncs()
 
     def setAndUpdateValues(self, solution_next, IncShkDstn, LivPrb, DiscFac):